requirements.txt
fastapi>=0.115.14
uvicorn>=0.34.3
flask>=3.0.3
streamlit>=1.45.1
requests>=2.32.3
python-multipart>=0.0.22
pydantic>=2.11.4
pytest>=8.1.1
httpx>=0.28.1
pillow>=10.0.0
